drivers: dma_mcux_lpc: Replace sem with spinlock
Replace the otrig configuration semaphore with a spinlock.

This will allow the dma_config function to be called from an ISR
and it will no longer shift the burden of waiting to be able to configure
the dma otrigs to the caller of the function, since the driver
will just spin on the lock until it can configure them, instead of
returning an error.

(cherry picked from commit 0daad872fd9fad9000901516d842f17ed003a6c0)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 0daad872fd9fad9000901516d842f17ed003a6c0
Change-Id: Ib19696ca24c71ca93974aa734955bd8fd3f31c15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4585184
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>