net: dns_sd: Prevent dead code in query parsing
The number of buffer provided was verified in the final else block of a
long validation sequence. It would never be executed though, as one of
the conditions before would always evaluate to true.

As the number of buffers provided verification appears to be significant
in this case, as the buffers are referenced during other validations,
move this check at the beginning of the sequence instead. This also
eliminates the dead-code problem.

(cherry picked from commit 0ae7812f6bdb61eba9434b9fee18c7e9350c27e8)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: 0ae7812f6bdb61eba9434b9fee18c7e9350c27e8
Change-Id: Id44fd596cba80fc74b072c656edbff1bba2e7cb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4586373
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>