rp2040: Fix I2C frequency for all boards
The DW driver used by the RP2040 boards uses a Kconfig symbol
to determine the clock speed of the board. This value was
set incorrectly by the pico board to 100MHz (instead of 125).
For the other boards, it wasn't set at all.
This commit fixes the kconfig defaults. In the future, this
value should be taken directly from the devicetree. This is
not currently possible as there is no clock control driver
for the RP2040 (see #52901).

(cherry picked from commit 90079903bf738e84cd5b338324e291ef4b1c701a)

Original-Signed-off-by: Yonatan Schachter <yonatan.schachter@gmail.com>
GitOrigin-RevId: 90079903bf738e84cd5b338324e291ef4b1c701a
Change-Id: I6c79eb2eed84fb8f7aa2752b7eced4634ff1052a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4591723
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>