Bluetooth: OTS: Fix offset maximum range check
If offset is only 32-bit, then it can never be > UINT32_MAX,
so added another conditional that only if the type (off_t) is
larger than 32-bit, we perform the maximum value check.

off_t is not a standard type and thus the size of it is poorly
defined, and are just defined as a signed integer type.

(cherry picked from commit 9c13b85665e76daff9afaf0eeec2de4a99079623)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 9c13b85665e76daff9afaf0eeec2de4a99079623
Change-Id: I556ffaadef858563ec8b48fa2447504f6c82f3dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4591716
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>