Revert "drivers: serial: ns16550: Add support for Async APIs"
This reverts commit 2d03aaf99f326ec0c916da1ca7b8725c8b9fdbc8.

The async API for NS16550 is incomplete. We are near the next
release so it is better to revert it for now, and a proper
correct implementation can be done before next-next release.

Relates to #57103

(cherry picked from commit d88840a8aabd2a0d620a750397ad646c04874959)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: d88840a8aabd2a0d620a750397ad646c04874959
Change-Id: I33ba6d18812a53d512fba0e350b33e1643642fb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4614371
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>