cmake: Treat native libraries differently where needed
When building a native library there is a few options
we cannot pass to the compiler and linker,
including instructing them to genrate non PIE code
(as it is still to early to say that),
or garbage collect unused sections (as we are not yet doing
the final linking).

We also need to provide different
link options when building the elf for the DTS gen_handles
parsing (as that script requires a "final" executable elf)
than when we build the native library itself.

(cherry picked from commit feaf0ff5761b8f4deb1b8e93d4e760975ac28739)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: feaf0ff5761b8f4deb1b8e93d4e760975ac28739
Change-Id: Id9c7b18d24670e269cd3b7028d0377dfc80dda02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4668914
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>