samples/net: Add cellular modem sample
This commit adds a sample application which demonstrates how
to use the new driver and modules. The sample uses power
management to turn on the modem, uses network management
to wait for L4 connected, then uses DNS to get the IP
of the server running the python script found in the
server folder, which echoes back data recevied to it.
A packet containing psudo random data is then sent to
the server, which the echoes it back. To validate the
capability of the driver to restart the modem, the
modem is restarted, and the packet is sent again.

The server is hosted by linode, and uses the domain
name test-endpoint.com

(cherry picked from commit c0c9d6f7d65b56f709ae3c081cbdb32c61489e92)

Original-Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>
GitOrigin-RevId: c0c9d6f7d65b56f709ae3c081cbdb32c61489e92
Change-Id: I906b75c7c0e2b2e5fe971d84af352d3b7534917f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4830489
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>