net: lwm2m: Fix possible out-of-bound access when creating FW object
obj_inst_id should not be used directly to index the instance array, as
the instance ID is not tightly bound to the maximum instance count and
can exceed this value, causing out-of-bound access.

Therefore, perform some extra validation when choosing the array index
for the object instance to make sure we stay in the array bounds, or
return an error if there's no more room for more object instances, in a
similar way it's done for Security object.

(cherry picked from commit 479036165c780a9c5fad4829b33ac405399aefd9)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: 479036165c780a9c5fad4829b33ac405399aefd9
Change-Id: I562828ef6978c99a8533a3e57192f53240b12e3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4598334
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>