manifest: bsim: Do not import but replicate in main manifest
Due to a limitation in the west import feature,
a project cannot both have an import and be part of a group.
Moreover, when a project has an import and is not filtered
out, it is required for that project to be present
for most west commands to work.

As the bsim project is not filtered by default,
it causes trouble for users who never run a
west update but try to use west further.

To work around this issue, let's disable the import
in the bsim project, and instead replicate its content
inside Zephyr's main manifest.

Having a replica of the babblesim manifest content
is likely to cause some confusion in users,
wrt to which version of components they are using.
So whenever west supports both imports and groups, or another
simple and nicer way of working around this, it should be used.

(cherry picked from commit 9ce0d416f513b9108c13f07e4927c2400d2b6c72)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 9ce0d416f513b9108c13f07e4927c2400d2b6c72
Change-Id: I48cc27c60e7c1dc411b86d5da74c6a7b01194694
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4604571
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>