fs: littlefs: use uintptr_t instead of int for dev_id
dev_id is used as a pointer at the end of the call chain, so passing it
as an int makes no sense and can cause crashes if the pointer and int
types have different sizes. For example, if we have 64-bit pointers on
a board, the higher part of dev_id will be removed due to an type cast.

(cherry picked from commit d7241b0a2c2dca8807318e93958c4c93faa44589)

Original-Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>
Original-Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>
GitOrigin-RevId: d7241b0a2c2dca8807318e93958c4c93faa44589
Change-Id: I93074997a72fe42e863d4827d32610032316a39d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4758831
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>