sysbuild: Fix exporting BOARD.+ through sysbuild cache
Lately, sysbuild started filtering out variable names matching ^BOARD
during sysbuild cache file generation. As a result, one of the cache
variables that is no longer exported to other domains is BOARD_ROOT,
which breaks the recent support for out-of-tree boards in sysbuild.

Of those variables, the only one which is reasonable to filter out is
BOARD itself, because the purpose of that is to substitute a different
value for the one found in sysbuild's own CMake cache, which has the
board revision stripped out.

Update the relevant if-condition and use a single regex for brevity.

Fixes #59114.

(cherry picked from commit 87aa25e4ada12c3b65e455cc6e2fbd693be9c17c)

Original-Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>
GitOrigin-RevId: 87aa25e4ada12c3b65e455cc6e2fbd693be9c17c
Change-Id: I86adb6f060609d472808ad211f5363730be00fbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4608554
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>