drivers: sensor: ism330dhcx: fix pulse mode interrupts
On my end, the ISM330DHCX was stopping working after a few seconds.

After investigation, it seems that the function used to set the device
in pulse mode only works for special modes like tap and embedded functions
and not for data-ready as intended.

The data-ready was then in the default latched mode which does not work
sustainably with the rest of the driver logic. In this mode, the driver
can miss an interrupt and be forever waiting on a new data-ready pulse
which will never happen as the interrupt line is already active.

This calls the correct function to enable pulsed data-ready mode as
described in the datasheet section 9.7 COUNTER_BDR_REG1 (0Bh).

(cherry picked from commit 96a887809cbf52e5e2a25830f1f2b9358a8623b9)

Original-Signed-off-by: Francois Gervais <francoisgervais@gmail.com>
GitOrigin-RevId: 96a887809cbf52e5e2a25830f1f2b9358a8623b9
Change-Id: I0b67f83c1ab8206fef83627e413408e87c93edee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4830502
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>