tests: portability cmsis_rtosv2: Fix buffer aligment
The kernel requires the buffer to be word aligned.
Instead of assuming the word size is 32bits,
lets align the buffer to the size of a pointer,
which should match the word size (and which
is the check the kernel performs).

(cherry picked from commit 585063606a13437608a2e87f3602bf29ac3f8707)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 585063606a13437608a2e87f3602bf29ac3f8707
Change-Id: I21193a14a5ef19fab97b1bc5a5de93d663085299
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4631633
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>