drivers: dma_mcux_edma: add support dma driver for s32k344
On S32K344, the offset in memory map between each channel
is 0x4000 for most channels, but there is specific case is
between channel 11 and 12 which is 0x1D4000 instead. As a
consequence, 32 channels are divided to two parts: one
starts from channel 0 -> 11. The other is from channel 128
to 145. The channel gap is from 12 -> 127.

For user and data structures in shim driver, the channel's
value comes from 0 --> 31. Above constraint will be counted
when interact with the mcux sdk

Beside that, the DMAMUX register in this platform is very
specific, not in identical with DMAMUX channel, so shim
driver is updated to cover this case

(cherry picked from commit 8185faa0cb7c3a04081d8317dab2293568080714)

Original-Signed-off-by: Dat Nguyen Duy <dat.nguyenduy@nxp.com>
GitOrigin-RevId: 8185faa0cb7c3a04081d8317dab2293568080714
Change-Id: If37589da0f56993b018726e224d31815cf4a6759
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4900842
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>