Bluetooth: CSIP: Update CSIP duplicate rank search
In the CSIP Set Coordinator verify_members we verify that
all members have unique ranks.

This was done as part of the initial loop, but it is
more efficient to do this as a separate loop as we can reduce
the number of items to compare, as well as guarding this properly
by !zero_rank.

(cherry picked from commit 2b74450544abfaa73d7d4c7dde6b644e8b4b8ed3)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 2b74450544abfaa73d7d4c7dde6b644e8b4b8ed3
Change-Id: I2e514f33d85160a66581ef0a5cf313879602408d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4654455
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>