cache: Add note about buffer alignment and size
Cache operations that act on unaligned buffers (with respect to cache
line) or buffers that are not multiple of the cache line size are at
least dangerous with an high risk of data and memory corruption.

While we do not enforce a contract on the buffer characteristics between
users and APIs, we can at least add a note in the documentation
specifying an undefined behaviour when the passed buffer is unaligned or
wrongly sized.

(cherry picked from commit 0ace886359c622217d008b67eee02c9285cc11b3)

Original-Signed-off-by: Carlo Caione <ccaione@baylibre.com>
GitOrigin-RevId: 0ace886359c622217d008b67eee02c9285cc11b3
Change-Id: I647afbed89d0502481de648d172f1faa13720f34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4636347
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>