doc: Update requirements-doc.txt to use docleaf
Instead of breathe.

I'm not sure of the best version comparison to use. There is no
intention to break the API for docleaf but it is also still young
in some ways and there might be adjustments to the configuration values
that would need to be reflected in the version number as it is still
pre-1.0 that would only be a minor version change.

(cherry picked from commit 30c3ce4a92c7ebcd7885ac5a01ddcfc0334fbad4)

Original-Signed-off-by: Michael Jones <m.pricejones@gmail.com>
GitOrigin-RevId: 30c3ce4a92c7ebcd7885ac5a01ddcfc0334fbad4
Change-Id: Ibed66f1023c6068ae705beb6f3c16cb4ea34bfcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4661362
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>