doc: Make sure message is a string before regexing it
There are times when the log message is an exception object and that
causes the regex to fail with a TypeError. Here we work around that
by converting the input to a string.

(cherry picked from commit 56d7afeac9408f216e5e785ee5c90b740293fed0)

Original-Signed-off-by: Michael Jones <m.pricejones@gmail.com>
GitOrigin-RevId: 56d7afeac9408f216e5e785ee5c90b740293fed0
Change-Id: If0bbb25f86ba989e87b8719f63adfc7e69c83cc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4661366
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>