doc: Add '--keep-going' flag to sphinx-build
So that we see all the potential errors in CI and not just the first
and can assess how much work there is to do.

This was introduced for debuggin but gmarull has suggested keeping it
in.

(cherry picked from commit 343f60c6314689b8942c446708706a5c8ea9fe13)

Original-Signed-off-by: Michael Jones <m.pricejones@gmail.com>
GitOrigin-RevId: 343f60c6314689b8942c446708706a5c8ea9fe13
Change-Id: Ic5a6d9daa18129568d5d9a16fa739315cbb9a46c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4661367
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>