boards: add vendor to board yaml
This is coming from devicetree and corrosponds to what we have in the
dts/bindings/vendor-prefixes.txt file.

This will allow for static filtering, especially with twister, i.e. no
need to build anything to know the vendor of a board

All of the vendor data was extracted automatically from the devicetree,
so some platforms might not have the right vendor or no vendor at all
right now, we need to fix some of the DTS information or do this
manually to get this 100% correct. But we are close.

(cherry picked from commit c3827ec48e2acf5f4c6ec9a2c57e8e38b864b1c0)

Original-Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: c3827ec48e2acf5f4c6ec9a2c57e8e38b864b1c0
Change-Id: Ie07e8acf70c46f872250a9318cf3e189450a4961
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4887377
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>