drivers: hwspinlock: implement sqn hwspinlock driver
When we lock an hwspinlock, we must write the CPU identifier to
the hwspinlock register. If we want to unlock the locked hwspinlock,
we have to rewrite the same CPU identifier.

To define the CPU identifier, we use affinity 1 and affinity 2 fields
of the MPIDR register.

(cherry picked from commit f66b73197d4828cc55e363b5e78b6fdd17690f83)

Original-Signed-off-by: Aziz Idomar <aidomar@sequans.com>
GitOrigin-RevId: f66b73197d4828cc55e363b5e78b6fdd17690f83
Change-Id: I628b63c3b99b3f5a38d818cf302e3d3205bd8154
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4789647
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>