doc: Log unused log filter patterns
This allows us to identify any patterns that we can remove or that
we didn't realise are no longer in use. This might happen if issues
within doxygen or docleaf are resolved.

This allows us to remove the pattern:

  .*Duplicate C declaration.*\n.*'\.\. c:.*:: uint16_t id'.*

which does not match anything in the current set up.

We also split the filter patterns in known-warnings out into different
sections depending on their cause.

Also extend the pattern parser to ignore empty lines so that we can
have some formatting in the known-warnings file.

(cherry picked from commit 34e50037f0d94efce10770da5abcf48f7ba33906)

Original-Signed-off-by: Michael Jones <m.pricejones@gmail.com>
GitOrigin-RevId: 34e50037f0d94efce10770da5abcf48f7ba33906
Change-Id: I349b52bc3306defa852b650505a6b48fce606913
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4664199
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>